retrofit funcionando conectando con backend

+++ app/src/main/java/com/example/soundlink/core/domain/repository/UserRepository.kt
@@ -4,7 +4,7 @@ import com.example.soundlink.core.domain.model.User
 
 interface UserRepository {
     suspend fun login(email: String, password: String): Boolean
-    suspend fun register(name: String, email: String, password: String, age: Int): Boolean
+    suspend fun register(name: String, email: String, password: String, age: Long): Boolean
 
     suspend fun getCurrentUser(email: String): User?
 }
